deliverables/member1_phase2_clean_ascii/requirements.txt
torch>=2.2
torchvision>=0.17
pandas>=2.2
matplotlib>=3.8
tqdm>=4.66
Pillow>=10.0
